src/requirements.txt
#pip install -r requirements.txt
pika
python-dotenv